requirements.txt
httpx==0.28.1
python-dotenv==1.1.0
beautifulsoup4==4.13.4
openai==1.86.0
requests==2.32.3